New SFO features and UI fixes
- Show thumbnails in the offline view.
- SFO Contextual menu.
- Single and multiple remove from offline.
- Allow open offline files.

+++ MegaApp/common/Services/FileService.cs
@@ -45,7 +45,8 @@ namespace MegaApp.Services
 
         public static void DeleteFile(string path)
         {
-            File.Delete(path);
+            if(File.Exists(path))
+                File.Delete(path);            
         }
 
         public static void ClearFiles(IEnumerable<string> filesToDelete)
